+++ app/src/main/java/com/soop_assignment/app/domain/model/User.kt
@@ -3,9 +3,9 @@ package com.soop_assignment.app.domain.model
 data class User(
     val userImageUrl: String,
     val userName: String,
-    val followers: Int,
-    val following: Int,
+    val followers: Long,
+    val following: Long,
     val languages: String,
-    val repositories: Int,
+    val repositories: Long,
     val bio: String
 )
